bug: Fixed permanent environment check Moved CloudFormation to it's own nested stack
